+++ next-docs/public/examples/dropdownTW/__tests__/index.test.tsx
@@ -15,7 +15,7 @@ import Select from '../Select';
 import SelectStates from '../SelectStates';
 import TriggerElements from '../TriggerElements';
 
-const renderWithTheme = (component: JSX.Element) => (
+const renderWithLightTheme = (component: JSX.Element) => (
   <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
 );
 
@@ -23,34 +23,85 @@ const renderWithDarkTheme = (component: JSX.Element) => (
   <ThemeProvider theme={moonDesignDark}>{component}</ThemeProvider>
 );
 
-describe('Dropdown', () => {
-  it('renders default', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Default />));
+const renderWithRtl = (component: JSX.Element) => (
+  <div dir="rtl">{component}</div>
+);
+
+describe('Dropdown in Light Theme', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<Default />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
+  it('renders TriggerElements', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<TriggerElements />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders OptionsVariants', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<OptionsVariants />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders Select', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<Select />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders SelectStates', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<SelectStates />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});
 
-  it('renders default dark', () => {
+describe('Dropdown in Dark Theme', () => {
+  it('renders Default', () => {
     const testRenderer = renderer.create(renderWithDarkTheme(<Default />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
-
-  it('renders different options variant', () => {
-    const testRenderer = renderer.create(renderWithTheme(<OptionsVariants />));
+  it('renders TriggerElements', () => {
+    const testRenderer = renderer.create(
+      renderWithDarkTheme(<TriggerElements />)
+    );
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
-
-  it('renders select', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Select />));
+  it('renders OptionsVariants', () => {
+    const testRenderer = renderer.create(
+      renderWithDarkTheme(<OptionsVariants />)
+    );
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
-
-  it('renders select states', () => {
-    const testRenderer = renderer.create(renderWithTheme(<SelectStates />));
+  it('renders Select', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<Select />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
+  it('renders SelectStates', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<SelectStates />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});
 
-  it('renders different trigger elements', () => {
-    const testRenderer = renderer.create(renderWithTheme(<TriggerElements />));
+describe('Dropdown in RTL', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders TriggerElements', () => {
+    const testRenderer = renderer.create(renderWithRtl(<TriggerElements />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders OptionsVariants', () => {
+    const testRenderer = renderer.create(renderWithRtl(<OptionsVariants />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders Select', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Select />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders SelectStates', () => {
+    const testRenderer = renderer.create(renderWithRtl(<SelectStates />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 });