+++ next-docs/public/examples/menuItemTW/__tests__/index.test.tsx
@@ -3,7 +3,11 @@
  */
 
 import React from 'react';
-import { moonDesignLight, ThemeProvider } from '@heathmont/moon-themes';
+import {
+  moonDesignDark,
+  moonDesignLight,
+  ThemeProvider,
+} from '@heathmont/moon-themes';
 import renderer from 'react-test-renderer';
 import Default from '../Default';
 import AsLink from '../AsLink';
@@ -15,53 +19,159 @@ import Radio from '../Radio';
 import WithIcon from '../WithIcon';
 import WithMeta from '../WithMeta';
 
-const renderWithTheme = (component: JSX.Element) => (
+const renderWithLightTheme = (component: JSX.Element) => (
   <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
 );
 
-describe('Loader', () => {
-  it('renders default', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Default />));
+const renderWithDarkTheme = (component: JSX.Element) => (
+  <ThemeProvider theme={moonDesignDark}>{component}</ThemeProvider>
+);
+
+const renderWithRtl = (component: JSX.Element) => (
+  <div dir="rtl">{component}</div>
+);
+
+describe('Loader in Light Theme', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<Default />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders as link', () => {
-    const testRenderer = renderer.create(renderWithTheme(<AsLink />));
+  it('renders AsLink', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<AsLink />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders with checkbox', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Checkbox />));
+  it('renders Checkbox', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<Checkbox />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders with radio', () => {
-    const testRenderer = renderer.create(renderWithTheme(<Radio />));
+  it('renders Radio', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<Radio />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders expand and collapse', () => {
-    const testRenderer = renderer.create(renderWithTheme(<ExpandCollapse />));
+  it('renders ExpandCollapse', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<ExpandCollapse />)
+    );
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders multi line', () => {
-    const testRenderer = renderer.create(renderWithTheme(<MultiLine />));
+  it('renders MultiLine', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<MultiLine />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders multi title', () => {
-    const testRenderer = renderer.create(renderWithTheme(<MultiTitle />));
+  it('renders MultiTitle', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<MultiTitle />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders with icon', () => {
-    const testRenderer = renderer.create(renderWithTheme(<WithIcon />));
+  it('renders WithIcon', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<WithIcon />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 
-  it('renders with meta', () => {
-    const testRenderer = renderer.create(renderWithTheme(<WithMeta />));
+  it('renders WithMeta', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<WithMeta />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});
+
+describe('Loader in Dark Theme', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders AsLink', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<AsLink />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders Checkbox', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<Checkbox />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders Radio', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<Radio />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders ExpandCollapse', () => {
+    const testRenderer = renderer.create(
+      renderWithDarkTheme(<ExpandCollapse />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders MultiLine', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<MultiLine />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders MultiTitle', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<MultiTitle />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders WithIcon', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<WithIcon />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders WithMeta', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<WithMeta />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});
+
+describe('Loader in RTL', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders AsLink', () => {
+    const testRenderer = renderer.create(renderWithRtl(<AsLink />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders Checkbox', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Checkbox />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders Radio', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Radio />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders ExpandCollapse', () => {
+    const testRenderer = renderer.create(renderWithRtl(<ExpandCollapse />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders MultiLine', () => {
+    const testRenderer = renderer.create(renderWithRtl(<MultiLine />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders MultiTitle', () => {
+    const testRenderer = renderer.create(renderWithRtl(<MultiTitle />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders WithIcon', () => {
+    const testRenderer = renderer.create(renderWithRtl(<WithIcon />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+
+  it('renders WithMeta', () => {
+    const testRenderer = renderer.create(renderWithRtl(<WithMeta />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
 });