+++ workspaces/tables/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @heathmont/moon-table-tw
 
+## 9.12.2
+
+### Patch Changes
+
+- Table fix
+
+## 9.12.1
+
+### Patch Changes
+
+- Deploy fixes
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.12.1
+
 ## 9.12.0
 
 ### Minor Changes